# src/server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: SecureTag AI API
  description: |
    # SecureTag AI - Documentación Técnica y Guía de Integración

    Bienvenido a la documentación oficial de SecureTag AI. Esta guía detalla la arquitectura, capacidades, seguridad y proceso de integración de nuestra plataforma de Auditoría de Código de Nueva Generación (Next-Gen SAST).

    ---

    ## 1. Visión General

    SecureTag AI redefine el análisis estático de seguridad (SAST) combinando herramientas de escaneo líderes en la industria con un motor de Inteligencia Artificial Cognitiva.

    A diferencia de los escáneres tradicionales, SecureTag AI:
    1.  **Detecta** vulnerabilidades en el código fuente.
    2.  **Analiza** el contexto arquitectónico mediante IA.
    3.  **Valida** los hallazgos para reducir falsos positivos.
    4.  **Recomienda** soluciones precisas y accionables.

    ---

    ## 2. Arquitectura y Seguridad

    Su instancia de SecureTag AI opera bajo una arquitectura de aislamiento estricto ("Tenant Isolation"), garantizando que sus datos y código permanezcan protegidos.

    ### 2.1 Componentes del Sistema
    *   **SecureTag API**: Puerta de enlace segura para la gestión de auditorías.
    *   **Analysis Engine**: Orquestador de herramientas SAST con capacidad de "Resilient Scanning" para grandes repositorios.
    *   **AI Security Core**: Modelo cognitivo (`securetag-v1`) optimizado para ciberseguridad, alojado en infraestructura privada (GPU).
...

...

...

...

...

...

...

...

...

    ---

...

...

...

...

...

...

    ---
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/auth/sync:
    post:
      summary: Synchronize User Identity (Frontend <-> Backend)
      description: |
        Internal endpoint used by Frontend Server to register or sync a user with the Backend Core.
        If user exists, returns their tenant/user IDs.
        If user is new, creates a Personal Tenant and Admin User.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                source:
                  type: string
      responses:
        200:
          description: User synced successfully (Existing)
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  tenant_id: { type: string, format: uuid }
                  role: { type: string }
                  is_new: { type: boolean }
        201:
          description: User created successfully (New)
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  tenant_id: { type: string, format: uuid }
                  role: { type: string }
                  is_new: { type: boolean }
        400:
          description: Invalid Input
        401:
          description: Unauthorized (Missing System Secret)
...